Fixes for cache abstraction
1. Correctly set ttl for redis keys
2. Correctly set default value for local caffeine cache
3. Sensible defaults to cache builder set in constructor

+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/security/DefaultSecurityService.java
@@ -172,7 +172,7 @@ public class DefaultSecurityService
             return;
         }
         
-        Integer attempts = (Integer)userFailedLoginAttemptCache.get( username ).get();
+        Integer attempts = userFailedLoginAttemptCache.get( username ).get();
         
         attempts++;
         
@@ -198,7 +198,7 @@ public class DefaultSecurityService
             return false;
         }
         
-        return (Integer) userFailedLoginAttemptCache.get( username ).get() > LOGIN_MAX_FAILED_ATTEMPTS;
+        return userFailedLoginAttemptCache.get( username ).get() > LOGIN_MAX_FAILED_ATTEMPTS;
     }
     
     private boolean isBlockFailedLogins()
